Some changes

+++ modules/common-ui-components/src/main/java/com/ankitsuda/rebound/ui/components/ExerciseItem.kt
@@ -16,12 +16,10 @@ package com.ankitsuda.rebound.ui.components
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.ankitsuda.rebound.ui.theme.LocalThemeState
@@ -32,7 +30,7 @@ fun ExerciseItem(
     modifier: Modifier = Modifier,
     name: String,
     muscle: String,
-    totalLogs: Int,
+    totalLogs: Long,
     onClick: () -> Unit
 ) {
     Box(modifier = modifier) {
